+++ src/index.js
@@ -6,6 +6,8 @@ class Jimple {
         this.tagmap = new Map();
         this.shared = new Map();
         this.frozen = new Set();
+
+        Object.freeze(this);
     }
 
     /**
@@ -254,6 +256,38 @@ class Jimple {
         }
 
         return this.values.get(name);
+    }
+
+    proxify() {
+        try {
+            require.resolve("harmony-reflect");
+        } catch (err) {
+            throw new Error("You should install harmony-reflect module and use the --harmony_proxies flag to use this feature");
+        }
+
+        require("harmony-reflect");
+
+        let handler = {
+            get: (target, name) => {
+                return name in target ? target[name] : target.get(name);
+            },
+
+            set: (target, name, value, receiver) => {
+                if (name in target) {
+                    throw new Error(`Cannot define a service named ${name}`);
+                }
+
+                target.share(name, value);
+
+                return true;
+            },
+
+            has: (target, name) => target.exists(name),
+
+            deleteProperty: () => false
+        };
+
+        return new Proxy(this, handler);
     }
 }